Switch from com.bc.zarr:jzarr to dev.zarr:jzarr
Also merge ZarrPixelBufferTest.java logic into ZarrPixelBufferTest
