+++ build.gradle
@@ -26,9 +26,6 @@ repositories {
         url 'https://artifacts.glencoesoftware.com/artifactory/scijava-thirdparty'
     }
     maven {
-        url 'https://artifacts.glencoesoftware.com/artifactory/jzarr-releases/'
-    }
-    maven {
         url 'https://repo.glencoesoftware.com/repository/bioformats2raw2ometiff/'
     }
 }
@@ -36,7 +33,7 @@ repositories {
 dependencies {
     api 'org.openmicroscopy:omero-blitz:5.6.2'
     implementation 'com.github.ben-manes.caffeine:caffeine:3.1.8'
-    implementation 'com.bc.zarr:jzarr:0.3.5'
+    implementation 'dev.zarr:jzarr:0.4.2'
     implementation 'org.lasersonlab:s3fs:2.2.3'
     implementation 'com.amazonaws:aws-java-sdk-s3:1.12.659'
